REFACTOR: refactoring MainScreen and BookedScreen
