+++ src/components/PostList/index.tsx
@@ -2,15 +2,28 @@ import React, { useCallback } from 'react';
 import { StyleSheet, View, FlatList } from 'react-native';
 
 import Post from 'components/Post';
-import { IPost } from 'interfaces';
+import { IPost, AppNavigationParamList } from 'interfaces';
+import { StackNavigationProp } from '@react-navigation/stack';
 
 type PostListProps = {
   data: any;
-  openItem: (post: IPost) => void;
+  navigation: StackNavigationProp<
+    AppNavigationParamList,
+    'MainScreen' | 'BookedScreen'
+  >;
 };
 
-const PostList: React.FC<PostListProps> = ({ data, openItem }) => {
+const PostList: React.FC<PostListProps> = ({ data, navigation }) => {
   const keyExtractor = useCallback((post: IPost) => post.id.toString(), []);
+
+  const openItem = (post: IPost) => {
+    navigation.navigate('PostScreen', {
+      postId: post.id,
+      postDescription: post.text,
+      postDate: post.date,
+      isBooked: post.booked,
+    });
+  };
 
   return (
     <View style={styles.container}>
