+++ src/screens/MainScreen/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { StackNavigationProp } from '@react-navigation/stack';
 
 import DATA from 'data';
-import { IPost, AppNavigationParamList } from 'interfaces';
+import { AppNavigationParamList } from 'interfaces';
 import PostList from 'components/PostList';
 
 type MainScreenProps = {
@@ -10,16 +10,7 @@ type MainScreenProps = {
 };
 
 const MainScreen: React.FC<MainScreenProps> = ({ navigation }) => {
-  const openItem = (post: IPost) => {
-    navigation.navigate('PostScreen', {
-      postId: post.id,
-      postDescription: post.text,
-      postDate: post.date,
-      isBooked: post.booked,
-    });
-  };
-
-  return <PostList data={DATA} openItem={openItem} />;
+  return <PostList data={DATA} navigation={navigation} />;
 };
 
 export default MainScreen;
